requirements.txt
Django==3.1.8
psycopg2-binary==2.8.6
djangorestframework==3.12.4
Pillow==8.2.0
django-markupfield==2.*
environs[django]==9.3.2
rollbar==0.15.2
django-mptt==0.12.0
django-taggit==1.3.*
django-taggit-helpers==0.1.4
more-itertools==8.7.0
django-filter==2.4.0
django-import-export==2.5.0
gTTS==2.2.2
sqlalchemy==1.4.7
beautifulsoup4[lxml]==4.9.3
tqdm==4.60.0
gTTS-token==1.1.4